Actor Refactor(WIP) + Small changes
Still need to add perpetual actions. Changed Java source version to Java 14(yay arrow switch cases C:). Going to work on perpetual and path porting rn if I have time
