+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Auton/Actor.java
@@ -1,282 +1,151 @@
 package org.firstinspires.ftc.teamcode.Auton;
 
-
-import static org.firstinspires.ftc.teamcode.Auton.Config.tolerance;
-import static org.firstinspires.ftc.teamcode.Auton.Config.toleranceH;
-
-import android.widget.ArrayAdapter;
-
-import com.acmerobotics.dashboard.FtcDashboard;
-
 import com.arcrobotics.ftclib.geometry.Pose2d;
-import com.arcrobotics.ftclib.geometry.Rotation2d;
-import com.arcrobotics.ftclib.geometry.Translation2d;
 import com.qualcomm.robotcore.hardware.HardwareMap;
-import com.qualcomm.robotcore.util.ElapsedTime;
+import com.qualcomm.robotcore.util.Range;
 
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.function.Supplier;
-
-import org.checkerframework.checker.units.qual.A;
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.teamcode.Control.Movement;
 import org.firstinspires.ftc.teamcode.Controllers.PID;
-import org.firstinspires.ftc.teamcode.Subsystems.Claw;
-import org.firstinspires.ftc.teamcode.Subsystems.Lift;
 import org.firstinspires.ftc.teamcode.Subsystems.Robot;
-import org.firstinspires.ftc.teamcode.Control.Movement;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.utils;
 
-public class Actor {
-    HardwareMap hw;
-    Telemetry tm;
-    Robot robot;
-    SampleMecanumDrive rrDrive;
+import java.lang.reflect.Array;
+import java.util.ArrayList;
 
-    ArrayList<ArrayList<Action>> actions = new ArrayList<>();
-    ArrayList<Action> perpetualActions = new ArrayList<>();
-    ElapsedTime timer = null;
-    int indexOn = 0;
-    double maxTime;
-    public Actor(HardwareMap hw, Telemetry tm, Robot robot, SampleMecanumDrive rrDrive, double maxTimePerAction)
+public class Actor {
+    private final HardwareMap hw;
+    private final Telemetry tm;
+    private final Robot robot;
+    private final SampleMecanumDrive rrDrive;
+    private final Movement movement;
+    private final double defaultTimeout;
+
+
+    public Actor(HardwareMap hw, Telemetry tm, Robot robot, SampleMecanumDrive rrDrive, Movement movement, double defaultTimeout)
     {
         this.hw = hw;
         this.tm = tm;
         this.robot = robot;
         this.rrDrive = rrDrive;
-        this.maxTime = maxTimePerAction;
+        this.movement = movement;
+        this.defaultTimeout = defaultTimeout;
     }
-    public void add(ActionInput a)
-    {
-        this.add(a, false, false);
+
+    private ArrayList<ArrayList<Action>> actions = new ArrayList<>();
+    public void add(Action action, boolean parallel) {
+        if (parallel) {
+            actions.get(actions.size() - 1).add(action);
+        } else {
+            actions.add(new ArrayList<>());
+            actions.get(actions.size() - 1).add(action);
+        }
     }
-    public void add(ActionInput a, boolean inParallel)
-    {
-        this.add(a, inParallel, false);
-    }
-    public void add(ActionInput a, boolean inParallel, boolean perpetual)//inParallel = in parallel with the last listed action. don't use inparallel if its the first action
-    {
-        if(!inParallel) // if separate steps
-        {
-            this.actions.add(new ArrayList<>()); // add completely new step to actions
-        }
 
-
-        Action x = null;
-        if(a.type == ActionInput.inputType.LIFT)
-        {
-            x = new LiftAction(robot.lift, a.args[0], a.args[1]);
-        }
-        else if(a.type == ActionInput.inputType.CLAW)
-        {
-            x = new ClawAction(robot.claw, a.args[0], a.args[1]);
-        }
-        else if(a.type == ActionInput.inputType.MOVEMENT)
-        {
-            x = new MovementAction(new Pose2d(new Translation2d(a.args[0], a.args[1]), new Rotation2d(a.args[2])),
-                    robot,
-                    rrDrive,
-                    a.args[3]/100.0);
-        }
-
-
-
-        if(x != null)
-        {
-            ArrayList<Action> temp = new ArrayList<>();//creates new arraylist to update perpetualactions
-            for(int i = 0; i<this.perpetualActions.size(); i++)
-            {
-                if(!perpetualActions.get(i).getClass().getName().equals(x.getClass().getName())) {
-                    temp.add(perpetualActions.get(i));//add already perpetual actions and does overriding
-                }
-            }
-            if(perpetual)
-            {
-                temp.add(x);//will add this as a new perpetual action if necessary
-            }
-            this.perpetualActions = temp;
-        }
-
-        if(!inParallel)
-        {
-            for(Action b : perpetualActions)
-            {
-                if(b != x)
-                {
-                    this.actions.get(actions.size()-1).add(b);
-                }
+    public void run() {
+        ArrayList<Action> step = actions.get(0);
+        boolean stepDone = true;
+        for (Action action : step) {
+            // if the action is not done we run it and set stepDone to false so we don't move onto the next step until all actions are completed
+            if (!action.isDone(hw, tm, robot, rrDrive, movement)) {
+                stepDone = false;
+                action.run(hw, tm, robot, rrDrive, movement);
             }
         }
-        this.actions.get(actions.size()-1).add(x);
-    }
-    public void clearPerpetual()
-    {
-        this.perpetualActions = new ArrayList<>();
-    }
-    public void add(ConfigChange c, boolean inParallel)
-    {
-        if(!inParallel)
-        {
-            this.actions.add(new ArrayList<>());
-        }
-        this.actions.get(actions.size()-1).add(c);
-    }
-    public double actingFor()
-    {
-        if(timer!=null)
-        {
-            return timer.milliseconds();
-        }
-        return -1;
-    }
-
-
-
-    public void act()
-    {
-        boolean finished = true;
-        if(indexOn == actions.size())
-        {
-            return;
-        }
-        for(Action a : this.actions.get(indexOn))
-        {
-            finished = finished && a.act(); // before was &
-            if(timer == null)
-            {
-                timer = new ElapsedTime();
-            }
-        }
-        if(finished || timer.milliseconds() > maxTime)
-        {
-            indexOn++;
-            timer = new ElapsedTime();
-        }
-    }
-
-    static class ActionInput
-    {
-        public enum inputType{LIFT, CLAW, MOVEMENT};
-        inputType type;
-        int[] args;
-        public ActionInput(inputType type, int[] args)
-        {
-            this.type=type;
-            this.args=args;
-        }
+        if (stepDone) {
+            actions.remove(0);
+        } // we move onto the next step
     }
 }
-abstract class Action
-{
-    public abstract boolean act();
+
+abstract class Action {
+    public abstract void run(HardwareMap hw, Telemetry tm, Robot robot, SampleMecanumDrive rrDrive, Movement movement);
+    public abstract boolean isDone(HardwareMap hw, Telemetry tm, Robot robot, SampleMecanumDrive rrDrive, Movement movement);
 }
 
-class ConfigChange extends Action
-{
-    Runnable run;
-    public ConfigChange(Runnable run)
-    {
-        this.run = run;
+class ClawAction extends Action {
+    enum ClawStates{
+        topOpen,
+        topClosed,
+        bottomOpen,
+        bottomClosed
     }
-    public boolean act()
-    {
-        run.run();
-        return true;
-    }
-}
-class LiftAction extends Action{
 
-    Lift lift;
-    int liftTo;
-    int liftPower;
-    public LiftAction(Lift lift, int liftTo, int liftPower)
-    {
-        this.lift = lift;
-        this.liftTo = liftTo;
-        this.liftPower = liftPower;
-
+    private final ClawStates[] states;
+    public ClawAction(ClawStates[] states) {
+        this.states = states;
     }
     @Override
-    public boolean act() {
-        lift.liftToPos(liftTo, liftPower);
-        if(Math.abs(lift.rightLift.getCurrentPosition() - liftTo) < 15)
-        {
-            return true;
+    public void run(HardwareMap hw, Telemetry tm, Robot robot, SampleMecanumDrive rrDrive, Movement movement) {
+        for (ClawStates state : states) {
+            switch (state) {
+                case topOpen -> robot.claw.topServo.setPower(Config.topServoOpen);
+                case topClosed -> robot.claw.topServo.setPower(Config.topServoClose);
+                case bottomOpen -> robot.claw.bottomServo.setPower(Config.bottomServoOpen);
+                case bottomClosed -> robot.claw.bottomServo.setPower(Config.bottomServoClose);
+            }
         }
-        return false;
     }
-}
-
-class ClawAction extends Action
-{
-    Claw claw;
-    int bottom, top;
-    public ClawAction(Claw claw, int bottom, int top)
-    {
-        this.claw = claw;
-        this.bottom = bottom;
-        this.top = top;
-    }
-
     @Override
-    public boolean act() {
-        double bottomPower = this.bottom==0 ? Config.bottomServoClose : Config.bottomServoOpen;
-        double topPower = this.top==0 ? Config.topServoClose : Config.topServoOpen;
-        claw.setPower(bottomPower, topPower);
+    public boolean isDone(HardwareMap hw, Telemetry tm, Robot robot, SampleMecanumDrive rrDrive, Movement movement) {
+        // Claw is a bit scuffed, because of the way I setup parallel actions, since isDone is always true(as we don't know the servo positon since we aren't using axons)
+        // the run function would otherwise never be called, so we call it ourselves here before returning true
+        this.run(hw, tm, robot, rrDrive, movement);
         return true;
     }
 }
 
-class MovementAction extends Action
-{
-    Pose2d target;
-    SampleMecanumDrive rrDrive;
-    PID driveXPID, driveYPID, headingPID;
-    Robot robot;
-    double power;
-    public MovementAction(Pose2d destination, Robot robot, SampleMecanumDrive rrDrive, double power)
-    {
-        this.target = destination;
-        this.robot=robot;
-        this.rrDrive=rrDrive;
-        this.driveXPID = new PID(Config.translationP,Config.translationI,Config.translationD);
-        this.driveYPID = new PID(Config.translationP,Config.translationI,Config.translationD);
-        this.headingPID = new PID(Config.rotationP,Config.rotationI,Config.rotationD);
-        this.power=power;
+class LiftAction extends Action {
 
+    private final int height;
+    private final double power;
+    private final AUTON_RED_NEAR.CAFPid pid = new AUTON_RED_NEAR.CAFPid(new PID.Config(Config.liftP, Config.liftI, Config.liftD));
+
+    public LiftAction(int height, double power) {
+        this.height = height;
+        this.power = power;
     }
+
     @Override
-    public boolean act() {
-        rrDrive.update();
+    public void run(HardwareMap hw, Telemetry tm, Robot robot, SampleMecanumDrive rrDrive, Movement movement) {
+        // not gonna use liftToPos so this way we can allow for custom pid per section in the future
+        double p = Range.clip(pid.calc(height, height - robot.lift.leftLift.getCurrentPosition()),-power, power);
+        robot.lift.setLiftPower(p + Config.gravity); // gravity is F in a traditional PIDF controller
+    }
+
+    @Override
+    public boolean isDone(HardwareMap hw, Telemetry tm, Robot robot, SampleMecanumDrive rrDrive, Movement movement) {
+        return height - robot.lift.leftLift.getCurrentPosition() < Config.liftTolerance;
+    }
+}
+
+// Movement Action, called MvntAction to persevere the same 10 letter length to make it more readable
+class MvntAction extends Action {
+    private final Pose2d target;
+    private final double maxPower;
+    public MvntAction(Pose2d target, double maxPower) {
+        this.target = target;
+        this.maxPower = maxPower;
+    }
+
+    public MvntAction(Pose2d target) {
+        this.target = target;
+        this.maxPower = Config.powerMultiplier;
+    }
+
+    // TODO add an option for just driving for time(for example into the wall)
+    // TODO add a correct way for setting maxPower(maybe we do it similar to how lift will have its own pid)
+    @Override
+    public void run(HardwareMap hw, Telemetry tm, Robot robot, SampleMecanumDrive rrDrive, Movement movement) {
+        movement.move(target, maxPower);
+    }
+
+    @Override
+    public boolean isDone(HardwareMap hw, Telemetry tm, Robot robot, SampleMecanumDrive rrDrive, Movement movement) {
+        // same as in movement.move, just inverted to be isDone instead of continueNextLoop
+        double tolerance = Config.tolerance;
         Pose2d pose = rrDrive.getPose();
-
-        double x = driveXPID.getValue(target.getX() - pose.getX());
-        double y = driveYPID.getValue(target.getY() - pose.getY());
-        double rx = headingPID.getValue(utils.angleDifference(target.getRotation().getDegrees(), Math.toDegrees(pose.getHeading())));
-        double botHeading = -pose.getRotation().getRadians();    //i won't change it cause it seems to work but y?
-        // just smt that happened when trying to get two wheel working, fixed in getPose()
-
-        double rotX = x * Math.cos(-botHeading) - y * Math.sin(-botHeading);
-        double rotY = x * Math.sin(-botHeading) + y * Math.cos(-botHeading);
-        rotX = rotX * Config.XMULTI;
-        double denominator = Math.max(Math.abs(rotY) + Math.abs(rotX) + Math.abs(rx), 1);
-        double frontLeftPower = (rotY + rotX + rx) / denominator;
-        double backLeftPower = (rotY - rotX + rx) / denominator;
-        double frontRightPower = (rotY - rotX - rx) / denominator;
-        double backRightPower = (rotY + rotX - rx) / denominator;
-        //dont spam telemetry on the actual driver hub, use dash if u want all this data
-
-        robot.drive.setDrivePowers(frontLeftPower*power, frontRightPower*power, backLeftPower*power, backRightPower*power);
-        //returns if we're there for the outside loop. can easily change to &&'s(which I recommend)
-
-        return Math.abs(target.getX() - pose.getX()) > tolerance || Math.abs(target.getY() - pose.getY()) > tolerance || Math.abs(utils.angleDifference(target.getRotation().getDegrees(), pose.getRotation().getDegrees())) > toleranceH;
+        return !(Math.abs(target.getX() - pose.getX()) > tolerance || Math.abs(target.getY() - pose.getY()) > tolerance || Math.abs(utils.angleDifference(target.getRotation().getDegrees(), pose.getRotation().getDegrees())) > toleranceH);
     }
-
-
 }
-
-
-
