+++ build.common.gradle
@@ -107,7 +107,7 @@ android {
     }
 
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_14
         targetCompatibility JavaVersion.VERSION_1_8
     }
